Move to search.h all the prototypes of non-static search.c functions
