+++ fontforge/python.c
@@ -80,6 +80,7 @@ extern int old_sfnt_flags;
 #include "scripting.h"
 #include "scriptfuncs.h"
 #include "scstyles.h"
+#include "search.h"
 #include "ffpython.h"
 
 #include <math.h>
