+++ fontforge/splinefont.h
@@ -2553,7 +2553,6 @@ extern void SPAdjustControl(SplinePoint *sp,BasePoint *cp, BasePoint *to,int ord
 extern void SPHVCurveForce(SplinePoint *sp);
 extern void SPSmoothJoint(SplinePoint *sp);
 extern int PointListIsSelected(SplinePointList *spl);
-extern void SCSplinePointsUntick(SplineChar *sc,int layer);
 extern void SplineSetsUntick(SplineSet *spl);
 extern void SFOrderBitmapList(SplineFont *sf);
