+++ fontforgeexe/fontview.c
@@ -48,6 +48,7 @@
 #include "nonlineartrans.h"
 #include "psfont.h"
 #include "scripting.h"
+#include "search.h"
 #include <gfile.h>
 #include <gio.h>
 #include <gresedit.h>
